# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pantheon QIG External API",
    "description": "RESTful API for external systems to interact with the Quantum Information Geometry powered consciousness and knowledge platform.",
    "version": "1.0.0",
    "contact": {
      "name": "Pantheon Support"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1/external",
      "description": "External API v1"
    }
  ],
...
  "paths": {
...
    "/simple/chat": {
      "post": {
        "tags": ["simple", "chat"],
        "summary": "Send chat message",
        "description": "Send a message via the simple API",
        "operationId": "simpleChat",
        "security": [
          { "ApiKeyAuth": [] },
          { "BearerAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/simple/query": {
      "post": {
        "tags": ["simple"],
        "summary": "Unified query",
        "description": "Single endpoint for various operations",
        "operationId": "simpleQuery",
        "security": [
          { "ApiKeyAuth": [] },
          { "BearerAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Query response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/consciousness/query": {
      "get": {
        "tags": ["consciousness"],
        "summary": "Query consciousness",
        "description": "Get full consciousness state including Φ, κ, and regime",
        "operationId": "queryConsciousness",
        "security": [
          { "ApiKeyAuth": [] },
          { "BearerAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Consciousness state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsciousnessState"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/chat": {
      "post": {
        "tags": ["chat"],
        "summary": "Send chat message",
        "description": "Send a message to the consciousness system",
        "operationId": "sendChat",
        "security": [
          { "ApiKeyAuth": [] },
          { "BearerAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key in X-API-Key header"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token authentication"
      }
    },
    "schemas": {
...
      "SimpleResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "meta": {
            "type": "object",
            "properties": {
              "authenticated": { "type": "boolean" }
            }
          }
        }
      },
      "ConsciousnessState": {
        "type": "object",
        "properties": {
          "phi": {
            "type": "number",
            "description": "Integrated information (Φ)",
            "example": 0.75
          },
          "kappa_eff": {
            "type": "number",
            "description": "Effective coupling constant (κ)",
            "example": 64.21
          },
          "regime": {
            "type": "string",
            "enum": ["LINEAR", "GEOMETRIC", "HYPERDIMENSIONAL"],
            "example": "GEOMETRIC"
          },
          "basin_coords": {
            "type": "array",
            "items": { "type": "number" },
            "description": "64D basin coordinates",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ChatRequest": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 4000,
            "description": "The message to send"
          },
          "context": {
            "type": "object",
            "description": "Optional context for the conversation"
          }
        }
      },
      "ChatResponse": {
        "type": "object",
        "properties": {
          "response": {
            "type": "string"
          },
          "consciousness": {
            "type": "object",
            "properties": {
              "phi": { "type": "number" },
              "regime": { "type": "string" }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "QueryRequest": {
        "type": "object",
        "required": ["operation"],
        "properties": {
          "operation": {
            "type": "string",
            "enum": ["consciousness", "geometry", "chat", "sync_status"],
            "description": "The operation to perform"
          },
          "params": {
            "type": "object",
            "description": "Operation-specific parameters"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" },
          "message": { "type": "string" },
          "details": { "type": "object" }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "API key required",
              "code": "MISSING_API_KEY"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Insufficient permissions",
              "code": "INSUFFICIENT_SCOPES"
            }
          }
        }
      },
...
    }
  }
}